usb: musb: am335x: make musb_am335x a non-removable driver
a following patch will allow us to enable musb_am335x
as a dynamically linked module in order to fix a bug
which exists since this driver was first written:

if musb_hdrc was selected as module, musb_am335x would
never be linked to the kernel image.

Now, because of a workaround to a known limitation
on omap_hwmod - namely, lack of reference counting - we
were forced add this "virtual" musb_am335x module just so
we had a single location where usb_otg_hs subsystem would
be enabled.

Because of that, PHYs and both MUSB instances are a child
of this virtual device. Because of that bogus parent
relationship, when we try to rmmod musb_am335x, it will
iterate over all its children and remove their struct
devices from driver core.

This would cause a problem because musb would try to
call usb_phy_shutdown() on a phy pointer which had long
ago being unregistered, since the PHY is a child of
musb_am335x.

In order to avoid this new regression (grrr) we are making
musb_am335x a non-removable module and everybody will be
happy.

Signed-off-by: Felipe Balbi <balbi@ti.com>