fbdev/da8xx-fb: fix race issues with palette loading
When the da8xx-fb sets up a new config, it first uses the DMA to load a
palette to the hardware and then sets up the DMA to send frame data.
This happens asynchronously, without locking, and causes race issues.
Furthermore, the whole DMA setup is not done if fbconsole is not
enabled.

This patch changes the palette loading to be synchronous, as it removes
any race issues and loading the palette should be a small operation. The
new function, lcd_load_palette, sets the DMA for palette loading, waits
for it to complete, and then sets up the DMA for frame data.

Also, the lcd_init function is changed to explicitly call
lcd_load_palette, so that everything is set up correctly without needing
fbconsole (or some other component) to activate the config.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>