phy: omap-usb2: don't suspend if already runtime suspended
When trying to echo mem > /sys/power/state when the USB2
PHY was already runtime suspended, we would see a WARN()
complaining about unbalanced clk_disable().

This was pretty obviously wrong in the PHY driver where
it would allow the PHY to be suspended after it was already
runtime_suspended.

Tested-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>