CLK/DTS: dra7-atl clock driver architecture change
To be able to integrate the ATL provided clocks to the clock tree we need
to change the driver architecture.
For seamless integration two types of DT binding will be needed:
- DT clock nodes to represent the ATL clocks
- binding for the ATL IP itself which is going to handle the hw configuration

With this change pm_runtime can handle the ATL clock on demand of it's use
and all the IP which needs ATL clock can be sure that it is enabled for
them.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>