OMAPDSS: SiI9022: disable D3 power state
The current driver sets the SiI9022 to D3 power state when disabling the
chip on module remove. To exit D3, the reset gpio must be used.
Unfortunately we don't have control for the reset GPIO.

To make the driver more usable as modules, comment the code out that
sets the chip to D3.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>