v4l: ti-vpe: Allocate vfd through video_device_alloc()
The video_device struct is currently a part of the driver data struct vpe_dev.
A vpe_dev instance is allocated by the driver, and the memory for the vfd is a
part of this struct.

The v4l2 core, however, manages the removal of the vfd region, through the
video_device's .release() op, and tries to free it seperately. This causes
memory corruption, and leads to issues when we try to re-insert the vpe module.

Use vide_device_alloc() v4l2-device helper, and let the v4l2 core manage the
allocation and free-ing of the video_device struct memory.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>