ARM: AM43x: Use gptimer as clocksource
The SyncTimer in AM43x seems to be clocked using an inaccuarte 32k
clock (CLK_32KHZ) derived from PER DPLL, causing system time to go
slowly (~10% deviation).

Use gptimer as clocksource instead, as is done in the case of AM335x
(which does not have a SyncTimer). With this, system time keeping works
accurately.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>