OMAPDSS: dpi: adopt pinctrl support
Update omapdss/dpi driver to set the state of the pins to:
- "default on enable
- "sleep" on disable

The disable and enable functions are called from a PM hook within
omapdss so the sleep and default states will be used with suspend
and resume, respectively.

By optionally putting the pins into sleep state in the suspend callback
we can accomplish two things.
- minimize current leakage from pins and thus save power,
- prevent the IP from driving pins output in an uncontrolled manner,
which may happen if the power domain drops the domain regulator.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>