ARM: OMAP2+: omap_hwmod: Add context ops to am43xx soc_ops
AM43XX contains PRM_*_CONTEXT registers like OMAP4 did to track loss
of context in modules so utilize the same get_context_loss_count soc_op
present for OMAP4 but define a custom update context loss function for
am43xx. Because we do not define context_offs for each hwmod, introduce
a new flag to indicate when we have so we only bother to update when
we know that it is valid.

Certain drivers use the context loss count to determine if context was
truly lost during a suspend operation to decide whether or not a
context restore is truly necessary. Without this, some drivers may
choose not to restore context when they have actually lost it and fail
to properly resume.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>