input: touchscreen: pixcir: also trigger IRQ on rising edge
We want to make sure we never miss penup/down events
and the easiest way is to also trigger IRQ on the rising
edge.

This will make sure that when we touch or release the
touchscreen surface, an IRQ will be fired and our ISR
handler will be called, thus avoiding missed events.

Fixes #D-01570 - Cap touch screen issue

Tested-by: Franklin Cooper Jr <fcooper@ti.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>