gpu: host1x: Provide a proper struct bus_type
Previously the struct bus_type exported by the host1x infrastructure was
only a very basic skeleton. Turn that implementation into a more full-
fledged bus to support proper probe ordering and power management.

Note that the bus infrastructure needs to be available before any of the
drivers can be registered, so the bus needs to be registered before the
host1x module. Otherwise drivers could be registered before the module
is loaded and trigger a BUG_ON() in driver_register(). To ensure the bus
infrastructure is always there, always build the code into the kernel
when enabled and register it with a postcore initcall.

Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Mark Zhang <markz@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>