gpio: lib-sysfs: Add 'wakeup' attribute
Add an attribute 'wakeup' to the GPIO sysfs interface which
allows marking/unmarking a GPIO as wake IRQ.

The file 'wakeup' is created in each exported GPIOs directory,
if an IRQ is associated with that GPIO and the irqchip
implements set_wake().

Writing 'enabled' to that file will enable wake for that GPIO,
while writing 'disabled' will disable wake.

Reading that file will return either 'disabled' or 'enabled'
depening on the currently set flag for the GPIO's IRQ.

v3:
 - add documentation
v2:
 - fix error path to unlock mutex before return

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Reviewed-by: Alexandre Courbot <acourbot@nvidia.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>