MODSIGN: /proc/keys is not unconditionally available
Documentation/module-signing.txt file is referring to /proc/keys file in
order to view all keys contained in the kernel's keyring. That file is not
universally avialble when CONFIG_KEYS is enabled, which is confusing. The
fact that the option needed for this procfs interface to exist contains
"_DEBUG_" in its name makes it even more confusing. Document this fact.

Signed-off-by: Jiri Kosina <jkosina@suse.cz>