oom: don't count on mm-less current process
out_of_memory() doesn't trigger the OOM killer if the current task is
already exiting or it has fatal signals pending, and gives the task access
to memory reserves instead.  However, doing so is wrong if out_of_memory()
is called by an allocation (e.g.  from exit_task_work()) after the current
task has already released its memory and cleared TIF_MEMDIE at exit_mm().
If we again set TIF_MEMDIE to post-exit_mm() current task, the OOM killer
will be blocked by the task sitting in the final schedule() waiting for
its parent to reap it.  It will trigger an OOM livelock if its parent is
unable to reap it due to doing an allocation and waiting for the OOM
killer to kill it.

Signed-off-by: Tetsuo Handa <penguin-kernel@I-love.SAKURA.ne.jp>
Acked-by: Michal Hocko <mhocko@suse.cz>
Cc: David Rientjes <rientjes@google.com>
Cc: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>