eventfs: avoid unused variable warning
An optimization patch from Chris Mason causes build warnings about an
unused variable.

The patch that broke this is currently in the akpm-current series, so this
fixup can be folded into the original patch.  I was expecting multiple
people to send a patch for this, so I waited a bit at first.

Fixes: 567162b87a5c5f  ("eventfd: don't take the spinlock in eventfd_poll")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Chris Mason <clm@fb.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>