ARM: error on over-sized XIP images
While looking at some kallsyms failures from a configuration from Arnd,
it became apparent that it was possible the linked kernel to be invalid
despite a successful link due to sections occupying the same ELF VMA
addresses, but different ELF LMA addresses.

In this case, the linker does not error, but the result is that the data
section overlaps the rodata section (or possibly other sections.)

Rather than allowing the link to silently succeed, we add an assertion
check for XIP kernels to detect this condition, and force a failure
should it occur.

Acked-by: Nicolas Pitre <nico@linaro.org>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>