powerpc/dma: Support 32-bit coherent mask with 64-bit dma_mask
Some drivers try to use a 64-bit dma_mask and a smaller (32-bit typically)
coherent dma mask.

We don't currently support that on platforms that do direct DMA
because:

 - We use the generic dma_set_coherent_mask()

 - It will use dma_supported() with the provided mask

 - Our imlpementation of the latter in
   dma_set_our dma_direct_dma_supported() will fail if the mask
   passed isn't big enough to cover all of memory (+ DMA offset)

This fixes it by making dma_direct_dma_supported() use the ZONE_DMA32
limit when it's enabled, and removing the ifdef CONFIG_FSL_SOC around
the code in dma_direct_alloc_coherent() that supports allocating
from the 32-bit zone. (Nowadays, most drivers have been fixed).

This carries an additional change which is to generalize the test
against memblock_end_of_DRAM() + "dma_offset" to 32-bit platforms,
instead of just unconditionally returning 1 for these. This should
work since the rest of the code in the direct DMA ops assume this
is true anyway.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>